separated phase module into package
